+++ test/Main.hs
@@ -6,7 +6,9 @@ import Test.Markdown (markdownTests)
 import Test.Notebook (notebookTests)
 import Test.Parser (parseTests)
 import Test.Render (renderTests)
+import Test.Repl (replTests)
 import Test.Run (runTests)
+import Test.Version (versionTests)
 
 main :: IO ()
 main =
@@ -17,5 +19,7 @@ main =
             , renderTests
             , markdownTests
             , notebookTests
+            , replTests
             , runTests
+            , versionTests
             ]